#2145 moved org settings back to preferences and removed page titles
